add : 데이터를 조작하는 InputUtils 클래스 생성
