+++ src/main/java/lotto/view/input/InputCheck.java
@@ -1,20 +1,7 @@
 package lotto.view.input;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-
 public class InputCheck {
-    public static List<Integer> parseIntegerList(String text) {
-        List<String> list = Arrays.asList(text.split(","));
-        List<Integer> numbers = new ArrayList<>();
-        for (String value : list) {
-            numbers.add(parseInteger(value));
-        }
-        return numbers;
-    }
-
-    private static int parseInteger(String value) throws NumberFormatException {
+    public static int parseInteger(String value) throws NumberFormatException {
         return Integer.parseInt(value);
     }
 }
